requirements.txt
# Core ML and Deep Learning
torch>=1.12.0
torchvision>=0.13.0
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Computer Vision and Image Processing
opencv-python>=4.6.0
Pillow>=9.2.0
albumentations>=1.3.0
imageio>=2.22.0

# Point Cloud Processing
open3d>=0.16.0
pyntcloud>=0.3.1

# Data Processing and Visualization
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.10.0

# Deep Learning Utilities
timm>=0.6.0
segmentation-models-pytorch>=0.3.0
ultralytics>=8.0.0

# Web Framework and API
fastapi>=0.85.0
uvicorn>=0.18.0
streamlit>=1.15.0
gradio>=3.4.0

# Utilities
tqdm>=4.64.0
wandb>=0.13.0
tensorboard>=2.10.0
hydra-core>=1.2.0
omegaconf>=2.2.0

# Data Loading and Augmentation
torch-audiomentations>=0.11.0
kornia>=0.6.0

# Metrics and Evaluation
torchmetrics>=0.10.0
pycocotools>=2.0.6

# Deployment
gunicorn>=20.1.0
docker>=6.0.0

# Research and Experiment Tracking
mlflow>=1.28.0
optuna>=3.0.0

# Additional utilities
h5py>=3.7.0
zarr>=2.13.0
joblib>=1.2.0 